backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openpyxl==3.1.2
pandas==2.1.3
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
